abstraction, annotations were added, refactoring was performed
